+++ py_helpers/getThermostatVariables.py
@@ -28,9 +28,6 @@ if device_mac == "Not_Set":
 
 
 try:
-
-
-
     thermostat_ = client.thermostats.info(device_mac=device_mac)
     from wyze_sdk.models.devices import ThermostatSystemMode
     outputDict = {